+++ CHANGELOG.md
@@ -17,6 +17,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Fixed
 
+- Added ON DELETE to `most_recent_crawl` table, such that it can handle previously-crawled but now-dead instances.
+
 ## [2.7.0 - 2018-08-18]
 
 ### Added